subgraph/ownercontract/tests/owner.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { Address } from "@graphprotocol/graph-ts"
import { OwnerSet } from "../generated/schema"
import { OwnerSet as OwnerSetEvent } from "../generated/Owner/Owner"
import { handleOwnerSet } from "../src/owner"
import { createOwnerSetEvent } from "./owner-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let oldOwner = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newOwner = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newOwnerSetEvent = createOwnerSetEvent(oldOwner, newOwner)
    handleOwnerSet(newOwnerSetEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("OwnerSet created and stored", () => {
    assert.entityCount("OwnerSet", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "OwnerSet",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "oldOwner",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "OwnerSet",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "newOwner",
      "0x0000000000000000000000000000000000000001"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
